Add files via upload
During Covid-19 pandemic, I got an helpful idea to kids who wants to learn but due to pandemic they didn't do so. So i decided to create an online platform to help kids to learn their school courses online.
Basically this is just a design not a complete project because i have front end skills and not backend. 
So this is an incomplete project but later will complete this.

Hope you get some idea what about this website is.